Always remove override file before building a new one

+++ hooks/commands/run.sh
@@ -11,12 +11,12 @@ cleanup() {
 # clean up docker containers on EXIT
 trap cleanup EXIT
 
+test -f "$override_file" && rm "$override_file"
+
 if build_image=$(get_prebuilt_image_from_metadata "$service_name") ; then
   echo "~~~ :docker: Creating a modified Docker Compose config"
   build_image_override_file "$service_name" "$build_image" \
     > "$override_file"
-elif [ -f "$override_file" ] ; then
-  rm "$override_file"
 fi
 
 echo "+++ :docker: Running command in Docker Compose service: $service_name"
@@ -27,11 +27,11 @@ echo "+++ :docker: Running command in Docker Compose service: $service_name"
 #   docker-compose run "app" go test
 
 if [[ -f "$override_file" ]]; then
-  run_docker_compose -f "$override_file" pull "$service_name"
-  run_docker_compose -f "$override_file" run "$service_name" $BUILDKITE_COMMAND
+  run_docker_compose -f "$override_file" pull "$service_name" && \
+    run_docker_compose -f "$override_file" run "$service_name" $BUILDKITE_COMMAND
 else
-  run_docker_compose pull "$service_name"
-  run_docker_compose run "$service_name" $BUILDKITE_COMMAND
+  run_docker_compose pull "$service_name" && \
+    run_docker_compose run "$service_name" $BUILDKITE_COMMAND
 fi
 
 exitcode=$?
